+++ examples/liquidity/withdrawSavers.ts
@@ -2,7 +2,10 @@ import { Network } from '@xchainjs/xchain-client'
 import { Midgard, MidgardCache, MidgardQuery } from '@xchainjs/xchain-midgard-query'
 import { ThorchainAMM, Wallet } from '@xchainjs/xchain-thorchain-amm'
 import { SaversWithdraw, ThorchainCache, ThorchainQuery, Thornode } from '@xchainjs/xchain-thorchain-query'
-import { assetFromString } from '@xchainjs/xchain-util'
+import { assetFromString, register9Rheader } from '@xchainjs/xchain-util'
+import axios from 'axios'
+
+register9Rheader(axios)
 
 /**
  * Withdraw lp function