+++ packages/xchain-midgard-query/src/midgard-cache.ts
@@ -9,7 +9,7 @@ const MILLISECOND_CACHE_SAVERS = 5000
 const defaultMidgard = new Midgard()
 
 /**
- * This class manages retrieving information from Midgard API and cached it
+ * This class retrieves raw information from Midgard API and cached it
  */
 export class MidgardCache {
   readonly midgard: Midgard
@@ -18,7 +18,7 @@ export class MidgardCache {
   /**
    * Constructor to create a ThorchainCache
    *
-   * @param midgard - an instance of the midgard API (could be pointing to stagenet,testnet,mainnet)
+   * @param {Midgard} midgard - an instance of the midgard API (could be pointing to stagenet,testnet,mainnet)
    * @returns MidgardCache
    */
   constructor(midgard = defaultMidgard) {
@@ -27,10 +27,21 @@ export class MidgardCache {
     this.cachedSavers = new CachedValue<SaverDetails>(() => this.midgard.getSavers(''), MILLISECOND_CACHE_SAVERS)
   }
 
+  /**
+   * Get info about existing pools in the protocol from Midgard API
+   *
+   * @returns {PoolDetail[]} Array of pools
+   */
   public async getPools(): Promise<PoolDetail[]> {
     return this.cachedPools.getValue()
   }
 
+  /**
+   * Returns the information of all the positions of a set of addresses in the THORChain savers product.
+   *
+   * @param {String} address Comma separated list of addresses
+   * @returns {SaverDetails} Array of savers positions
+   */
   public async getSavers(address: string): Promise<SaverDetails> {
     return this.cachedSavers.getValue(address)
   }