+++ packages/xchain-thorchain-query/src/utils/thornode.ts
@@ -41,7 +41,7 @@ export type ThornodeConfig = {
 const defaultThornodeConfig: Record<Network, ThornodeConfig> = {
   mainnet: {
     apiRetries: 3,
-    thornodeBaseUrls: [`https://thornode.ninerealms.com`, `https://thornode.thorswap.net`],
+    thornodeBaseUrls: [`https://thornode.ninerealms.com`],
   },
   stagenet: {
     apiRetries: 3,