+++ packages/xchain-utxo-providers/src/providers/blockcypher/blockcypher-data-provider.ts
@@ -201,7 +201,9 @@ export class BlockcypherProvider implements UtxoOnlineDataProvider {
       })
 
       //remove duplicates
-      const txs = response.txrefs.map((i) => i.tx_hash)
+      const txs: string[] = []
+      response.txrefs && txs.push(...response.txrefs.map((i) => i.tx_hash))
+      response.unconfirmed_txrefs && txs.push(...response.unconfirmed_txrefs.map((i) => i.tx_hash))
       const uniqTxs = [...new Set(txs)]
       const start = offset >= uniqTxs.length ? uniqTxs.length : offset
       const end = offset + limit >= uniqTxs.length ? uniqTxs.length : offset + limit