+++ packages/xchain-thorchain-amm/CHANGELOG.md
@@ -1,3 +1,16 @@
+# v0.7.4 (2023-10-09)
+
+## Update
+
+- Fix pending UTXOs bug
+- Increate default fee estimation
+
+# v0.7.3 (2023-10-06)
+
+## Update
+
+- Update deps for thorchain-query && thornode && midgard
+
 # v0.7.2 (2023-09-24)
 
 ## Update